Refactor exception package of mysql-dialect-exception module
