+++ shardingsphere-dialect-exception/shardingsphere-mysql-dialect-exception/src/main/java/org/apache/shardingsphere/dialect/mysql/exception/UnknownCollationException.java
@@ -15,14 +15,20 @@
  * limitations under the License.
  */
 
-package org.apache.shardingsphere.dialect.exception.connection;
+package org.apache.shardingsphere.dialect.mysql.exception;
 
+import lombok.Getter;
+import lombok.RequiredArgsConstructor;
 import org.apache.shardingsphere.dialect.exception.SQLDialectException;
 
 /**
- * Unsupported prepared statement exception.
+ * Unknown collation exception.
  */
-public final class UnsupportedPreparedStatementException extends SQLDialectException {
+@RequiredArgsConstructor
+@Getter
+public final class UnknownCollationException extends SQLDialectException {
     
-    private static final long serialVersionUID = 2185290474191327830L;
+    private static final long serialVersionUID = 6920150607711135228L;
+    
+    private final int collationId;
 }
